Add _repr static method to Mapper class and remove it from MixedIntegerProgramExplanation
